Move to relocated web classes
Refactor code to move from recently deprecated classes.

Closes gh-5822
